Move mutable::Mut to cell::RefCell
