+++ src/libstd/lib.rs
@@ -164,7 +164,6 @@ pub mod result;
 pub mod either;
 pub mod hashmap;
 pub mod cell;
-pub mod mutable;
 pub mod trie;
