+++ src/test/compile-fail/mut-ptr-cant-outlive-ref.rs
@@ -8,10 +8,10 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use std::mutable::Mut;
+use std::cell::RefCell;
 
 fn main() {
-    let m = Mut::new(0);
+    let m = RefCell::new(0);
     let p;
     {
         let b = m.borrow();
